+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-07-30 21:14:09
-// Commit: fix(theme): fix memory leak by cancelling theme subscription
+// Updated at: 2026-07-30 21:24:27
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '3266';
-  final String lastMessage = 'fix(theme): fix memory leak by cancelling theme subscription';
-  final DateTime timestamp = DateTime.parse('2026-07-30T21:14:09+05:00');
+  final String commitRef = '3268';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-07-30T21:24:27+05:00');
 }